+++ sync_viber/models/res_users_settings.py
@@ -1,10 +1,11 @@
 # Copyright 2022 Ivan Yelizariev <https://twitter.com/yelizariev>
+# License MIT (https://opensource.org/licenses/MIT).
 from odoo import fields, models
 
 
 class ResUsersSettings(models.Model):
     _inherit = "res.users.settings"
 
-    is_discuss_sidebar_category_whatsapp_chatapi_open = fields.Boolean(
-        "Is category WhatsApp open", default=True
+    is_discuss_sidebar_category_viber_open = fields.Boolean(
+        "Is category Viber open", default=True
     )