+++ sync_whatsapp/__manifest__.py
@@ -6,7 +6,7 @@
     "summary": """WhatsApp integration powered by Sync Studio and Chat API""",
     "category": "Marketing",
     "images": ["images/sync_whatsapp.jpg"],
-    "version": "15.0.1.0.0",
+    "version": "15.0.2.0.0",
     "application": False,
     "author": "IT Projects Labs, Eugene Molotov",
     "support": "help@itpp.dev",